+++ Source/Podio .NET/Models/Item.cs
@@ -8,7 +8,7 @@ namespace PodioAPI.Models
     public class Item
     {
         [JsonProperty("item_id")]
-        public int ItemId { get; set; }
+        public long ItemId { get; set; }
 
         [JsonProperty("external_id")]
         public string ExternalId { get; set; }
@@ -29,7 +29,7 @@ namespace PodioAPI.Models
         public string AppItemIdFormatted { get; set; }
 
         [JsonProperty("app_item_id")]
-        public int? AppItemId { get; set; }
+        public long? AppItemId { get; set; }
 
 
         [JsonProperty("created_by")]
@@ -71,7 +71,7 @@ namespace PodioAPI.Models
         public JArray References { get; set; }
 
         [JsonProperty("linked_account_id")]
-        public int? LinkedAccountId { get; set; }
+        public long? LinkedAccountId { get; set; }
 
         [JsonProperty("subscribed")]
         public bool? Subscribed { get; set; }
@@ -113,7 +113,7 @@ namespace PodioAPI.Models
         public Grant Grant { get; set; }
 
         [JsonProperty("file_ids")]
-        public List<int> FileIds { get; set; }
+        public List<long> FileIds { get; set; }
 
         [JsonProperty("tasks")]
         public List<Task> Tasks { get; set; }
@@ -140,14 +140,14 @@ namespace PodioAPI.Models
             return fieldInstance<T>(genericField, externalId);
         }
 
-        public T Field<T>(int fieldId)
+        public T Field<T>(long fieldId)
             where T : ItemField, new()
         {
             var genericField = this.Fields.Find(field => field.FieldId == fieldId);
             return fieldInstance<T>(genericField, null, fieldId);
         }
 
-        protected T fieldInstance<T>(ItemField genericField, string externalId = null, int? fieldId = null)
+        protected T fieldInstance<T>(ItemField genericField, string externalId = null, long? fieldId = null)
             where T : ItemField, new()
         {
             T specificField = new T();
