+++ Source/Podio .NET/Services/ApplicationService.cs
@@ -25,7 +25,7 @@ namespace PodioAPI.Services
         ///     value: full
         /// </param>
         /// <returns></returns>
-        public Application GetApp(int appId, string type = "full")
+        public Application GetApp(long appId, string type = "full")
         {
             string url = string.Format("/app/{0}", appId);
             var requestData = new Dictionary<string, string>
@@ -40,7 +40,7 @@ namespace PodioAPI.Services
         ///     <para>Podio API Reference: https://developers.podio.com/doc/applications/activate-app-43822 </para>
         /// </summary>
         /// <param name="appId"></param>
-        public void ActivateApp(int appId)
+        public void ActivateApp(long appId)
         {
             string url = string.Format("/app/{0}/activate", appId);
             _podio.Post<dynamic>(url);
@@ -52,7 +52,7 @@ namespace PodioAPI.Services
         ///     <para>Podio API Reference: https://developers.podio.com/doc/applications/deactivate-app-43821 </para>
         /// </summary>
         /// <param name="appId"></param>
-        public void DeactivateApp(int appId)
+        public void DeactivateApp(long appId)
         {
             string url = string.Format("/app/{0}/deactivate", appId);
             _podio.Post<dynamic>(url);
@@ -66,7 +66,7 @@ namespace PodioAPI.Services
         /// <param name="includeInactive"> True if inactive apps should be included, false otherwise.Default value: false </param>
         /// <param name="additionalAttributes">Additional attributes to include in query string</param>
         /// <returns></returns>
-        public List<Application> GetAppsBySpace(int spaceId, bool includeInactive = false, Dictionary<string, string> additionalAttributes = null)
+        public List<Application> GetAppsBySpace(long spaceId, bool includeInactive = false, Dictionary<string, string> additionalAttributes = null)
         {
             var requestData = new Dictionary<string, string>
             {
@@ -94,7 +94,7 @@ namespace PodioAPI.Services
         /// <param name="includeSpace"></param>
         /// <param name="appIds">todo: describe appIds parameter on GetFeatures</param>
         /// <returns></returns>
-        public List<string> GetFeatures(int[] appIds, bool includeSpace = false)
+        public List<string> GetFeatures(long[] appIds, bool includeSpace = false)
         {
             string appIdCSV = Utilities.ArrayToCSV(appIds);
             string url = "/app/features/";
@@ -113,7 +113,7 @@ namespace PodioAPI.Services
         /// <param name="appId"></param>
         /// <param name="fieldId"></param>
         /// <returns></returns>
-        public ApplicationField GetAppField(int appId, int fieldId)
+        public ApplicationField GetAppField(long appId, long fieldId)
         {
             string url = string.Format("/app/{0}/field/{1}", appId, fieldId);
             return _podio.Get<ApplicationField>(url);
@@ -124,14 +124,14 @@ namespace PodioAPI.Services
         /// </summary>
         /// <param name="query">Any search term to match.</param>
         /// <returns></returns>
-        public List<int> GetIconSuggestions(string query)
+        public List<long> GetIconSuggestions(string query)
         {
             string url = "/app/icon/search";
             var requestData = new Dictionary<string, string>
             {
                 {"query", query}
             };
-            return _podio.Get<List<int>>(url, requestData);
+            return _podio.Get<List<long>>(url, requestData);
         }
 
         /// <summary>
@@ -158,7 +158,7 @@ namespace PodioAPI.Services
         /// </summary>
         /// <param name="organizationId"></param>
         /// <returns></returns>
-        public List<Application> GetTopAppsForOrganization(int organizationId)
+        public List<Application> GetTopAppsForOrganization(long organizationId)
         {
             string url = string.Format("/app/org/{0}/top/", organizationId);
             return _podio.Get<List<Application>>(url);
@@ -170,7 +170,7 @@ namespace PodioAPI.Services
         ///     <para>Podio API Reference: https://developers.podio.com/doc/applications/delete-app-43693 </para>
         /// </summary>
         /// <param name="appId"></param>
-        public void DeleteApp(int appId)
+        public void DeleteApp(long appId)
         {
             string url = string.Format("/app/{0}", appId);
             _podio.Delete<dynamic>(url);
@@ -185,7 +185,7 @@ namespace PodioAPI.Services
         /// <param name="fieldId"></param>
         /// <param name="deleteValues"> True if the values for the fields should be deleted, false otherwise Default value: false </param>
         /// <returns></returns>
-        public ApplicationRevision DeleteAppField(int appId, int fieldId, bool deleteValues = false)
+        public ApplicationRevision DeleteAppField(long appId, long fieldId, bool deleteValues = false)
         {
             string url = string.Format("/app/{0}/field/{1}", appId, fieldId);
             dynamic requestData = new
@@ -216,7 +216,7 @@ namespace PodioAPI.Services
         ///     Order by the name of the app. Default value: score
         /// </param>
         /// <returns></returns>
-        public List<Application> GetAllUserApps(int[] excludeAppIds = null, int? referenceableInOrg = null,
+        public List<Application> GetAllUserApps(long[] excludeAppIds = null, long? referenceableInOrg = null,
             string right = null, string text = null, bool excludeDemo = false, int limit = 4, string order = "score")
         {
             string appIdCSV = Utilities.ArrayToCSV(excludeAppIds);
@@ -245,7 +245,7 @@ namespace PodioAPI.Services
         ///     value: full
         /// </param>
         /// <returns></returns>
-        public Application GetAppOnSpaceByURLLabel(int spaceId, string UrlLabel, string type = "full")
+        public Application GetAppOnSpaceByURLLabel(long spaceId, string UrlLabel, string type = "full")
         {
             string url = string.Format("/app/space/{0}/{1}", spaceId, UrlLabel);
             var requestData = new Dictionary<string, string>
@@ -279,7 +279,7 @@ namespace PodioAPI.Services
         /// <param name="appId"></param>
         /// <param name="externalId"></param>
         /// <returns></returns>
-        public ApplicationField GetAppField(int appId, string externalId)
+        public ApplicationField GetAppField(long appId, string externalId)
         {
             string url = string.Format("/app/{0}/field/{1}", appId, externalId);
             return _podio.Get<ApplicationField>(url);
@@ -291,9 +291,9 @@ namespace PodioAPI.Services
         /// </summary>
         /// <param name="spaceId"></param>
         /// <param name="appIds"></param>
-        public void UpdateAppOrder(int spaceId, List<int> appIds)
+        public void UpdateAppOrder(long spaceId, List<long> appIds)
         {
-            appIds = new List<int>();
+            appIds = new List<long>();
             string url = string.Format("/app/space/{0}/order", spaceId);
             _podio.Put<dynamic>(url, appIds);
         }
@@ -304,7 +304,7 @@ namespace PodioAPI.Services
         /// </summary>
         /// <param name="appId"></param>
         /// <param name="description"></param>
-        public void UpdateAppDescription(int appId, string description)
+        public void UpdateAppDescription(long appId, string description)
         {
             string url = string.Format("/app/{0}/description", appId);
             dynamic requestData = new
@@ -320,7 +320,7 @@ namespace PodioAPI.Services
         /// </summary>
         /// <param name="appId"></param>
         /// <param name="usage"></param>
-        public void UpdateAppUsageInstructions(int appId, string usage)
+        public void UpdateAppUsageInstructions(long appId, string usage)
         {
             string url = string.Format("/app/{0}/usage", appId);
             dynamic requestData = new
@@ -341,7 +341,7 @@ namespace PodioAPI.Services
         ///     integration, forms, items. If the value is not given all but the "items" feature will be selected.
         /// </param>
         /// <returns></returns>
-        public int InstallApp(int appId, int spaceId, string[] features = null)
+        public long InstallApp(long appId, long spaceId, string[] features = null)
         {
             features = features == null ? new string[] {"items"} : features;
             string url = string.Format("/app/{0}/install", appId);
@@ -351,7 +351,7 @@ namespace PodioAPI.Services
                 features = features
             };
             dynamic response = _podio.Post<dynamic>(url, requestData);
-            return (int) response["app_id"];
+            return (long) response["app_id"];
         }
 
         /// <summary>
@@ -360,7 +360,7 @@ namespace PodioAPI.Services
         /// </summary>
         /// <param name="appId"></param>
         /// <returns></returns>
-        public List<AppCalculation> GetCalculationsForApp(int appId)
+        public List<AppCalculation> GetCalculationsForApp(long appId)
         {
             string url = string.Format("/app/{0}/calculation/", appId);
             return _podio.Get<List<AppCalculation>>(url);
@@ -372,7 +372,7 @@ namespace PodioAPI.Services
         /// </summary>
         /// <param name="application"></param>
         /// <returns>The id of the newly created app</returns>
-        public int AddNewApp(Application application)
+        public long AddNewApp(Application application)
         {
             /*
                 Example Usage: Adding a new application with a text field and category field.
@@ -411,7 +411,7 @@ namespace PodioAPI.Services
                 Fields = application.Fields
             };
             dynamic response = _podio.Post<dynamic>(url, requestDate);
-            return (int) response["app_id"];
+            return (long) response["app_id"];
         }
 
         /// <summary>
@@ -420,7 +420,7 @@ namespace PodioAPI.Services
         /// </summary>
         /// <param name="appId"></param>
         /// <param name="field"></param>
-        public int AddNewAppField(int appId, ApplicationField field)
+        public long AddNewAppField(long appId, ApplicationField field)
         {
             /*
                Example Usage: Adding a new text field.
@@ -436,7 +436,7 @@ namespace PodioAPI.Services
             */
             string url = string.Format("/app/{0}/field/", appId);
             dynamic response = _podio.Post<dynamic>(url, field);
-            return (int) response["field_id"];
+            return (long) response["field_id"];
         }
 
         /// <summary>
@@ -524,7 +524,7 @@ namespace PodioAPI.Services
         /// <param name="appId"></param>
         /// <param name="fieldId"></param>
         /// <param name="config"></param>
-        public void UpdateAnAppField(int appId, int fieldId, FieldConfig config)
+        public void UpdateAnAppField(long appId, long fieldId, FieldConfig config)
         {
             string url = string.Format("/app/{0}/field/{1}", appId, fieldId);
             _podio.Put<dynamic>(url, config);
@@ -536,7 +536,7 @@ namespace PodioAPI.Services
         /// </summary>
         /// <param name="appId"></param>
         /// <returns></returns>
-        public ApplicationDependency GetAppDependencies(int appId)
+        public ApplicationDependency GetAppDependencies(long appId)
         {
             string url = string.Format("/app/{0}/dependencies/", appId);
             return _podio.Get<ApplicationDependency>(url);
@@ -548,7 +548,7 @@ namespace PodioAPI.Services
         /// </summary>
         /// <param name="spaceId"></param>
         /// <returns></returns>
-        public ApplicationDependency GetSpaceAppDependencies(int spaceId)
+        public ApplicationDependency GetSpaceAppDependencies(long spaceId)
         {
             string url = string.Format("/space/{0}/dependencies/", spaceId);
             return _podio.Get<ApplicationDependency>(url);
