+++ Source/Podio .NET/Services/CalendarService.cs
@@ -22,7 +22,7 @@ namespace PodioAPI.Services
         /// <param name="dateTo">The date to search to</param>
         /// <param name="priority">The minimum priority for the events to return Default value: 1</param>
         /// <returns></returns>
-        public IEnumerable<CalendarEvent> GetAppCalendar(int appId, DateTime dateFrom, DateTime dateTo,
+        public IEnumerable<CalendarEvent> GetAppCalendar(long appId, DateTime dateFrom, DateTime dateTo,
             int? priority = null)
         {
             string url = string.Format("/calendar/app/{0}/", appId);
@@ -64,7 +64,7 @@ namespace PodioAPI.Services
         /// <param name="userId"></param>
         /// <param name="token"></param>
         /// <returns></returns>
-        public string GetAppCalendarAsiCal(int appId, int userId, string token)
+        public string GetAppCalendarAsiCal(long appId, long userId, string token)
         {
             string url = string.Format("/calendar/app/{0}/ics/{1}/{2}/", appId, userId, token);
             var options = new Dictionary<string, bool>
@@ -83,7 +83,7 @@ namespace PodioAPI.Services
         /// <param name="userId"></param>
         /// <param name="token"></param>
         /// <returns></returns>
-        public string GetGlobalCalendarAsiCal(int userId, string token)
+        public string GetGlobalCalendarAsiCal(long userId, string token)
         {
             string url = string.Format("/calendar/ics/{0}/{1}/", userId, token);
             var options = new Dictionary<string, bool>
@@ -104,7 +104,7 @@ namespace PodioAPI.Services
         /// <param name="dateTo">The date to search to</param>
         /// <param name="priority">The minimum priority for the events to return Default value: 1</param>
         /// <returns></returns>
-        public IEnumerable<CalendarEvent> GetSpaceCalendar(int spaceId, DateTime dateFrom, DateTime dateTo,
+        public IEnumerable<CalendarEvent> GetSpaceCalendar(long spaceId, DateTime dateFrom, DateTime dateTo,
             int? priority = null)
         {
             string url = string.Format("/calendar/space/{0}/", spaceId);
@@ -126,7 +126,7 @@ namespace PodioAPI.Services
         /// <param name="userId"></param>
         /// <param name="token"></param>
         /// <returns></returns>
-        public string GetSpaceCalendarAsiCal(int spaceId, int userId, string token)
+        public string GetSpaceCalendarAsiCal(long spaceId, long userId, string token)
         {
             string url = string.Format("/calendar/space/{0}/ics/{1}/{2}/", spaceId, userId, token);
             var options = new Dictionary<string, bool>
@@ -145,7 +145,7 @@ namespace PodioAPI.Services
         /// <param name="userId"></param>
         /// <param name="token"></param>
         /// <returns></returns>
-        public string GetTaskCalendarAsiCal(int taskId)
+        public string GetTaskCalendarAsiCal(long taskId)
         {
             string url = string.Format("/calendar/task/{0}/ics/", taskId);
             var options = new Dictionary<string, bool>
@@ -202,7 +202,7 @@ namespace PodioAPI.Services
         /// <param name="limit">The maximum number of events to return in each group Default value: 5</param>
         /// <param name="priority">The minimum priority for the events to return Default value: 1</param>
         /// <returns></returns>
-        public CalendarSummary GetCalendarSummaryForSpace(int spaceId, int limit = 5, int priority = 1)
+        public CalendarSummary GetCalendarSummaryForSpace(long spaceId, int limit = 5, int priority = 1)
         {
             string url = string.Format("/calendar/space/{0}/summary", spaceId);
             var requestData = new Dictionary<string, string>
@@ -221,7 +221,7 @@ namespace PodioAPI.Services
         /// <param name="uid"></param>
         /// <param name="startDateTime"></param>
         /// <param name="endDateTime"></param>
-        public void UpdateCalendarEvent(int uid, DateTime startDateTime, DateTime endDateTime)
+        public void UpdateCalendarEvent(long uid, DateTime startDateTime, DateTime endDateTime)
         {
             string url = string.Format("/calendar/event/{0}", uid);
             dynamic requestData = new
