+++ Source/Podio .NET/Utils/Utilities.cs
@@ -2,7 +2,7 @@
 {
     internal class Utilities
     {
-        internal static string ArrayToCSV(int[] array, string splitter = ",")
+        internal static string ArrayToCSV(long[] array, string splitter = ",")
         {
             if (array != null && array.Length > 0)
                 return string.Join(splitter, array);
